fixed syncing and live updates issues, for notifications, and chats:
One error is that when admin goes to the dashboard and have not clicked on any chat, and some customer writes to a chat for a specific product, I dont get live updates or notifications, only once when the admin has selected a chat will that specific chat, even if admin is on another chat, get live updates and synced, non of the other chats which admin has not selected prior will get live updates.

Another bug is that, the unread messages gets incremented by 2, per message instead of 1.

The third bug is that when admin selects a chat that has unread messages, the ui updates and shows the messages as read by removing the notifications "bubble", and on page reloads, the notifications are gone as expected, but as soon as you select a chat notifications reappear always as 4 unread, so it seems like it does not get persisted to backend.